requirements.txt
httplib2
pypandoc
PyHamcrest
# coverage
# codacy-coverage
